Move relation serializers to serializer module.
