+++ api/resources_portal/serializers/__init__.py
@@ -3,3 +3,16 @@ resources_portal.views if it is necessary to prevent a cyclical
 dependency."""
 
 from resources_portal.serializers.material import MaterialDetailSerializer, MaterialSerializer
+from resources_portal.views.relation_serializers import (
+    AttachmentRelationSerializer,
+    FulfillmentNoteRelationSerializer,
+    GrantRelationSerializer,
+    MaterialRelationSerializer,
+    MaterialRequestIssueRelationSerializer,
+    MaterialRequestRelationSerializer,
+    MaterialShareEventsRelationSerializer,
+    OrganizationInvitationRelationSerializer,
+    OrganizationRelationSerializer,
+    ShippingRequirementRelationSerializer,
+    UserRelationSerializer,
+)
