+++ api/resources_portal/views/material_request_issue.py
@@ -7,7 +7,7 @@ from guardian.shortcuts import get_objects_for_user
 
 from resources_portal.models import MaterialRequest, MaterialRequestIssue, Organization
 from resources_portal.notifier import send_notifications
-from resources_portal.views.relation_serializers import MaterialRequestRelationSerializer
+from resources_portal.serializers import MaterialRequestRelationSerializer
 
 MATERIAL_ARCHIVED_ERROR = "Cannot open issues for requests whose materials have been archived."
 REQUEST_UNFULFILLED_ERROR = "Cannot open issues for requests which haven't been fulfilled."
