+++ api/resources_portal/views/user_organizations.py
@@ -6,7 +6,7 @@ from rest_framework.response import Response
 from rest_framework_extensions.mixins import NestedViewSetMixin
 
 from resources_portal.models import Organization, OrganizationUserAssociation, User
-from resources_portal.views.relation_serializers import OrganizationRelationSerializer
+from resources_portal.serializers import OrganizationRelationSerializer
 
 
 class UserOrganizationViewSet(NestedViewSetMixin, viewsets.ModelViewSet):
